+++ src/mesh_generator.rs
@@ -1,19 +1,23 @@
 use std::collections::HashMap;
 
-use glam::{Mat4, Vec3, Vec3A};
+use glam::{Mat3, Mat4, Vec2, Vec3, Vec3A};
 use ttf_parser::{Face, GlyphId};
 
 use crate::{
     error::MeshTextError,
     util::{
-        raster_to_mesh, raster_to_mesh_indexed, text_mesh_from_data, text_mesh_from_data_indexed,
-        GlyphOutlineBuilder,
+        mesh_to_flat_2d, mesh_to_indexed_flat_2d, raster_to_mesh, raster_to_mesh_indexed,
+        text_mesh_from_data, text_mesh_from_data_2d, text_mesh_from_data_indexed,
+        text_mesh_from_data_indexed_2d, GlyphOutlineBuilder,
     },
-    BoundingBox, Glyph, IndexedMeshText, MeshText, QualitySettings, TextSection, CacheType,
+    BoundingBox, CacheType, Glyph, IndexedMeshText, MeshText, QualitySettings, TextSection,
 };
 
 type Mesh = (Vec<Vec3A>, BoundingBox);
+type Mesh2D = (Vec<Vec2>, BoundingBox);
+
 type IndexedMesh = (Vec<u32>, Vec<Vec3A>, BoundingBox);
+type IndexedMesh2D = (Vec<u32>, Vec<Vec2>, BoundingBox);
 
 /// A [MeshGenerator] handles rasterizing individual glyphs.
 ///
@@ -80,7 +84,7 @@ impl MeshGenerator {
             cache: HashMap::new(),
             font: face,
             indexed_cache: HashMap::new(),
-            quality: quality,
+            quality,
             use_cache: true,
         }
     }
@@ -98,55 +102,59 @@ impl MeshGenerator {
             cache: HashMap::new(),
             font: face,
             indexed_cache: HashMap::new(),
-            quality: quality,
+            quality,
             use_cache: false,
         }
     }
 
     /// Fills the internal cache of a [MeshGenerator] with the given characters.
-    /// 
+    ///
     /// Arguments:
-    /// 
+    ///
     /// * `glyphs`: The glyphs that will be precached. Each character should appear exactly once.
-    /// * `flat`: Wether the flat or three-dimensional variant of the characters should be preloaded. 
+    /// * `flat`: Wether the flat or three-dimensional variant of the characters should be preloaded.
     /// If both variants should be precached this function must be called twice with this parameter set
     /// to `true` and `false`.
     /// * `cache`: An optional value that controls which cache will be filled. `None` means both caches will be filled.
-    /// 
+    ///
     /// Returns:
-    /// 
+    ///
     /// A [Result] indicating if the operation was successful.
-    /// 
+    ///
     /// # Example
-    /// 
+    ///
     /// ```rust
     /// use meshtext::MeshGenerator;
-    /// 
+    ///
     /// let font_data = include_bytes!("../assets/font/FiraMono-Regular.ttf");
     /// let mut generator = MeshGenerator::new(font_data);
-    /// 
+    ///
     /// let common = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789".to_string();
-    /// 
+    ///
     /// // Precache both flat and three-dimensional glyphs both for indexed and non-indexed meshes.
     /// generator.precache_glyphs(&common, false, None);
     /// generator.precache_glyphs(&common, true, None);
     /// ```
-    pub fn precache_glyphs(&mut self, glyphs: &String, flat: bool, cache: Option<CacheType>)  -> Result<(), Box<dyn MeshTextError>> {
+    pub fn precache_glyphs(
+        &mut self,
+        glyphs: &str,
+        flat: bool,
+        cache: Option<CacheType>,
+    ) -> Result<(), Box<dyn MeshTextError>> {
         if let Some(cache_type) = cache {
             match cache_type {
                 CacheType::Normal => {
                     for c in glyphs.chars() {
                         self.generate_glyph(c, flat, None)?;
                     }
-                },
+                }
                 CacheType::Indexed => {
                     for c in glyphs.chars() {
                         self.generate_glyph_indexed(c, flat, None)?;
                     }
-                },
+                }
             }
-        }
-        else {
+        } else {
             // If no type is set explicitely, both variants will be precached.
             for c in glyphs.chars() {
                 self.generate_glyph(c, flat, None)?;
@@ -159,7 +167,7 @@ impl MeshGenerator {
         Ok(())
     }
 
-    /// Generates the [BasicMeshText] of a single character with a custom transformation.
+    /// Generates the [MeshText] of a single character with a custom transformation.
     ///
     /// Arguments:
     ///
@@ -191,6 +199,38 @@ impl MeshGenerator {
         }
 
         Ok(text_mesh_from_data(mesh))
+    }
+
+    /// Generates the two-dimensional [MeshText] of a single character with a custom transformation.
+    ///
+    /// Arguments:
+    ///
+    /// * `glyph`: The character that should be converted to a mesh.
+    /// * `transform`: The 3x3 homogenous transformation matrix in column
+    /// major order that will be applied to this text.
+    ///
+    /// Returns:
+    ///
+    /// The desired two-dimensional [MeshText] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    fn generate_glyph_2d(
+        &mut self,
+        glyph: char,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<MeshText, Box<dyn MeshTextError>> {
+        let mesh = self.load_from_cache(glyph, true)?;
+        let mut mesh = mesh_to_flat_2d(mesh);
+
+        if let Some(value) = transform {
+            let transform = Mat3::from_cols_array(value);
+
+            for v in mesh.0.iter_mut() {
+                *v = transform.transform_point2(*v);
+            }
+            mesh.1.transform_2d(&transform);
+        }
+
+        Ok(text_mesh_from_data_2d(mesh))
     }
 
     /// Generates the [IndexedMeshText] of a single character with a custom transformation.
@@ -229,6 +269,41 @@ impl MeshGenerator {
         Ok(text_mesh_from_data_indexed(mesh))
     }
 
+    /// Generates the two-dimensional [IndexedMeshText] of a single character
+    /// with a custom transformation.
+    ///
+    /// This function generates a mesh with indices and vertices.
+    ///
+    /// Arguments:
+    ///
+    /// * `glyph`: The character that should be converted to a mesh.
+    /// * `transform`: The 3x3 homogenous transformation matrix in column
+    /// major order that will be applied to this text.
+    ///
+    /// Returns:
+    ///
+    /// The desired [IndexedMeshText] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    fn generate_glyph_indexed_2d(
+        &mut self,
+        glyph: char,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
+        let mesh = self.load_from_cache_indexed(glyph, true)?;
+        let mut mesh = mesh_to_indexed_flat_2d(mesh);
+
+        if let Some(value) = transform {
+            let transform = Mat3::from_cols_array(value);
+
+            for v in mesh.1.iter_mut() {
+                *v = transform.transform_point2(*v);
+            }
+            mesh.2.transform_2d(&transform);
+        }
+
+        Ok(text_mesh_from_data_indexed_2d(mesh))
+    }
+
     /// Generates the [Mesh] of a single character with a custom transformation given
     /// as a [Mat4].
     ///
@@ -254,7 +329,35 @@ impl MeshGenerator {
         for v in mesh.0.iter_mut() {
             *v = transform.transform_point3a(*v);
         }
-        mesh.1.transform(&transform);
+        mesh.1.transform(transform);
+
+        Ok(mesh)
+    }
+
+    /// Generates the [Mesh2D] of a single character with a custom transformation given
+    /// as a [Mat3].
+    ///
+    /// Arguments:
+    ///
+    /// * `glyph`: The character that should be converted to a mesh.
+    /// * `transform`: The 3x3 homogenous transformation matrix.
+    ///
+    /// Returns:
+    ///
+    /// The desired [Mesh2D] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    pub(crate) fn generate_glyph_with_glam_transform_2d(
+        &mut self,
+        glyph: char,
+        transform: &Mat3,
+    ) -> Result<Mesh2D, Box<dyn MeshTextError>> {
+        let mesh = self.load_from_cache(glyph, true)?;
+        let mut mesh = mesh_to_flat_2d(mesh);
+
+        for v in mesh.0.iter_mut() {
+            *v = transform.transform_point2(*v);
+        }
+        mesh.1.transform_2d(transform);
 
         Ok(mesh)
     }
@@ -286,7 +389,37 @@ impl MeshGenerator {
         for v in mesh.1.iter_mut() {
             *v = transform.transform_point3a(*v);
         }
-        mesh.2.transform(&transform);
+        mesh.2.transform(transform);
+
+        Ok(mesh)
+    }
+
+    /// Generates the [IndexedMesh2D] of a single character with a custom transformation given
+    /// as a [Mat3].
+    ///
+    /// This function handles indexed meshes.
+    ///
+    /// Arguments:
+    ///
+    /// * `glyph`: The character that should be converted to a mesh.
+    /// * `transform`: The 3x3 homogenous transformation matrix.
+    ///
+    /// Returns:
+    ///
+    /// The desired [IndexedMesh2D] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    pub(crate) fn generate_glyph_with_glam_transform_indexed_2d(
+        &mut self,
+        glyph: char,
+        transform: &Mat3,
+    ) -> Result<IndexedMesh2D, Box<dyn MeshTextError>> {
+        let mesh = self.load_from_cache_indexed(glyph, true)?;
+        let mut mesh = mesh_to_indexed_flat_2d(mesh);
+
+        for v in mesh.1.iter_mut() {
+            *v = transform.transform_point2(*v);
+        }
+        mesh.2.transform_2d(transform);
 
         Ok(mesh)
     }
@@ -307,7 +440,7 @@ impl MeshGenerator {
     /// process.
     fn generate_text_section(
         &mut self,
-        text: &String,
+        text: &str,
         flat: bool,
         transform: Option<&[f32; 16]>,
     ) -> Result<MeshText, Box<dyn MeshTextError>> {
@@ -363,6 +496,74 @@ impl MeshGenerator {
         Ok(text_mesh_from_data(mesh))
     }
 
+    /// Generates two-dimensional [MeshText] for a given text section.
+    ///
+    /// Arguments:
+    ///
+    /// * `text`: The text that should be converted to a mesh.
+    /// * `transform`: The optional 3x3 homogenous transformation matrix in column
+    /// major order that will be applied to this text.
+    ///
+    /// Returns:
+    ///
+    /// The desired [MeshText] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    fn generate_text_section_2d(
+        &mut self,
+        text: &str,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<MeshText, Box<dyn MeshTextError>> {
+        let base_transform = match transform {
+            Some(value) => Mat3::from_cols_array(value),
+            None => Mat3::IDENTITY,
+        };
+
+        let mut mesh = (Vec::new(), BoundingBox::empty());
+        let mut overall_advance = 0f32;
+
+        let mut chars_iter = text.chars();
+
+        // The first char will be handled differently if present.
+        if let Some(first_glyph) = chars_iter.next() {
+            let x_advance = self
+                .font
+                .glyph_hor_advance(self.glyph_id_of_char(first_glyph))
+                .unwrap_or(0) as f32
+                / self.font.height() as f32;
+
+            let transform =
+                base_transform * Mat3::from_translation(Vec2::new(overall_advance, 0f32));
+            let mut glyph_mesh =
+                self.generate_glyph_with_glam_transform_2d(first_glyph, &transform)?;
+
+            // Add vertices and replace bbox.
+            mesh.0.append(&mut glyph_mesh.0);
+            mesh = (mesh.0, glyph_mesh.1);
+
+            overall_advance += x_advance;
+        }
+
+        for glyph in chars_iter {
+            let x_advance = self
+                .font
+                .glyph_hor_advance(self.glyph_id_of_char(glyph))
+                .unwrap_or(0) as f32
+                / self.font.height() as f32;
+
+            let transform =
+                base_transform * Mat3::from_translation(Vec2::new(overall_advance, 0f32));
+            let mut glyph_mesh = self.generate_glyph_with_glam_transform_2d(glyph, &transform)?;
+
+            // Add vertices and adjust bbox.
+            mesh.0.append(&mut glyph_mesh.0);
+            mesh = (mesh.0, mesh.1.combine(&glyph_mesh.1));
+
+            overall_advance += x_advance;
+        }
+
+        Ok(text_mesh_from_data_2d(mesh))
+    }
+
     /// Generates the [MeshText] of a given text section.
     ///
     /// This function handles indexed meshes.
@@ -381,7 +582,7 @@ impl MeshGenerator {
     /// process.
     fn generate_text_section_indexed(
         &mut self,
-        text: &String,
+        text: &str,
         flat: bool,
         transform: Option<&[f32; 16]>,
     ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
@@ -453,6 +654,95 @@ impl MeshGenerator {
         }
 
         Ok(text_mesh_from_data_indexed(mesh))
+    }
+
+    /// Generates two-dimensional [MeshText] for a given text section.
+    ///
+    /// This function handles indexed meshes.
+    ///
+    /// Arguments:
+    ///
+    /// * `text`: The text that should be converted to a mesh.
+    /// * `transform`: The optional 3x3 homogenous transformation matrix in column
+    /// major order that will be applied to this text.
+    ///
+    /// Returns:
+    ///
+    /// The desired [MeshText] or an [MeshTextError] if anything went wrong in the
+    /// process.
+    fn generate_text_section_indexed_2d(
+        &mut self,
+        text: &str,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
+        let base_transform = match transform {
+            Some(value) => Mat3::from_cols_array(value),
+            None => Mat3::IDENTITY,
+        };
+
+        let mut mesh = (Vec::new(), Vec::new(), BoundingBox::empty());
+        let mut overall_advance = 0f32;
+        let mut index_offset = 0;
+
+        let mut chars_iter = text.chars();
+
+        // The first char will be handled differently if present.
+        if let Some(first_glyph) = chars_iter.next() {
+            let x_advance = self
+                .font
+                .glyph_hor_advance(self.glyph_id_of_char(first_glyph))
+                .unwrap_or(0) as f32
+                / self.font.height() as f32;
+
+            let transform =
+                base_transform * Mat3::from_translation(Vec2::new(overall_advance, 0f32));
+            let mut glyph_mesh =
+                self.generate_glyph_with_glam_transform_indexed_2d(first_glyph, &transform)?;
+
+            // Update index offset (note that glyph meshes can be empty).
+            if let Some(max) = glyph_mesh.0.iter().max() {
+                index_offset = *max + 1;
+            }
+
+            // Add vertices and replace bbox.
+            mesh.0.append(&mut glyph_mesh.0);
+            mesh.1.append(&mut glyph_mesh.1);
+            mesh = (mesh.0, mesh.1, glyph_mesh.2);
+
+            overall_advance += x_advance;
+        }
+
+        for glyph in chars_iter {
+            let x_advance = self
+                .font
+                .glyph_hor_advance(self.glyph_id_of_char(glyph))
+                .unwrap_or(0) as f32
+                / self.font.height() as f32;
+
+            let transform =
+                base_transform * Mat3::from_translation(Vec2::new(overall_advance, 0f32));
+            let mut glyph_mesh =
+                self.generate_glyph_with_glam_transform_indexed_2d(glyph, &transform)?;
+
+            // Offset indices.
+            for i in glyph_mesh.0.iter_mut() {
+                *i += index_offset;
+            }
+
+            // Update index offset (note that glyph meshes can be empty).
+            if let Some(max) = glyph_mesh.0.iter().max() {
+                index_offset = *max + 1;
+            }
+
+            // Add vertices and indices and adjust bbox.
+            mesh.0.append(&mut glyph_mesh.0);
+            mesh.1.append(&mut glyph_mesh.1);
+            mesh = (mesh.0, mesh.1, mesh.2.combine(&glyph_mesh.2));
+
+            overall_advance += x_advance;
+        }
+
+        Ok(text_mesh_from_data_indexed_2d(mesh))
     }
 
     /// Loads the given glyph from the cache or adds it.
@@ -692,22 +982,38 @@ impl MeshGenerator {
 impl TextSection<MeshText> for MeshGenerator {
     fn generate_section(
         &mut self,
-        text: &String,
+        text: &str,
         flat: bool,
         transform: Option<&[f32; 16]>,
     ) -> Result<MeshText, Box<dyn MeshTextError>> {
         self.generate_text_section(text, flat, transform)
+    }
+
+    fn generate_section_2d(
+        &mut self,
+        text: &str,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<MeshText, Box<dyn MeshTextError>> {
+        self.generate_text_section_2d(text, transform)
     }
 }
 
 impl TextSection<IndexedMeshText> for MeshGenerator {
     fn generate_section(
         &mut self,
-        text: &String,
+        text: &str,
         flat: bool,
         transform: Option<&[f32; 16]>,
     ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
         self.generate_text_section_indexed(text, flat, transform)
+    }
+
+    fn generate_section_2d(
+        &mut self,
+        text: &str,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
+        self.generate_text_section_indexed_2d(text, transform)
     }
 }
 
@@ -720,6 +1026,14 @@ impl Glyph<MeshText> for MeshGenerator {
     ) -> Result<MeshText, Box<dyn MeshTextError>> {
         self.generate_glyph(glyph, flat, transform)
     }
+
+    fn generate_glyph_2d(
+        &mut self,
+        glyph: char,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<MeshText, Box<dyn MeshTextError>> {
+        self.generate_glyph_2d(glyph, transform)
+    }
 }
 
 impl Glyph<IndexedMeshText> for MeshGenerator {
@@ -730,5 +1044,13 @@ impl Glyph<IndexedMeshText> for MeshGenerator {
         transform: Option<&[f32; 16]>,
     ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
         self.generate_glyph_indexed(glyph, flat, transform)
+    }
+
+    fn generate_glyph_2d(
+        &mut self,
+        glyph: char,
+        transform: Option<&[f32; 9]>,
+    ) -> Result<IndexedMeshText, Box<dyn MeshTextError>> {
+        self.generate_glyph_indexed_2d(glyph, transform)
     }
 }