+++ Changelog.md
@@ -4,3 +4,9 @@
 ## Version 0.1.1
 * Add precaching functionality
 ---
+## Version 0.2.0
+* Add functionality for generating 2D meshes
+* Add benchmarks using [criterion](https://crates.io/crates/criterion)
+* Shrink crate size by including only a subset of glyphs in the built-in font
+* Fix function signatures
+---